Build and link googlemaps geoservice
For iOS builds we need to link the service statically. To keep the
changes as few as possible on our side, we build the service statically
on all platforms and link it utilizing the same command.

In our main cpp file, we import the plugin with Q_IMPORT_PLUGIN() and
add a geoservice provider.